pyproject.toml
[project]
name = "restaurant-app"
version = "0.0.1"
dependencies = [
    "Flask>=3.0.3",
    "dependency-injector>=4.42.0b1",
    "pytest>=8.3.2",
    "sqlalchemy>=2.0.32",
    "Flask-Pydantic>=0.12.0",
    "pydantic>=2.8.2",
    "python-dotenv>=1.0.1",
    "PyYAML>=6.0.2"
]
requires-python = ">=3.12"
authors = [
    {name = "Henrik Binggl", email = "hbinggl.lba@fh-salzburg.ac.at"}
]
maintainers = [
    {name = "Henrik Binggl", email = "hbinggl.lba@fh-salzburg.ac.at"},
]
description = "A sample Flask application showcasing a restaurant reservation system"
readme = "Readme.md"
license = {file = "LICENSE"}
keywords = ["python", "flask", "restaurant", "software-design", "FH Salzburg"]
classifiers = [
    "Development Status :: 4 - Beta",
    "Programming Language :: Python"
]

[project.urls]
Homepage = "https://its-git.fh-salzburg.ac.at/hbinggl/software-design-python"
Repository = "https://its-git.fh-salzburg.ac.at/hbinggl/software-design-python"
